refactor(task_runner): extract spawn functions to scheduling.rs and split tests
Move log_task_failure_event, record_task_failure, spawn_task,
register_pending_task, spawn_preregistered_task,
spawn_task_with_worktree_detector, update_status, and mutate_and_persist
from mod.rs into scheduling.rs. Split the 1015-line test block from
mod.rs into two focused files: tests.rs (stream/state/interceptor tests)
and spawn_tests.rs (spawn lifecycle/phase/transient/count tests).

mod.rs is now a lean 30-line re-export facade; scheduling.rs is 745 lines
(under the 800-line U-16 limit). Both new test files are under 520 lines.
